Refactor into contract extensions file
